requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.13
openai==1.14.2
requests==2.31.0
python-dotenv==1.0.1
spotipy==2.23.0
yt-dlp==2024.4.24
pygame==2.5.2
anthropic==0.25.0
pydub==0.25.1
beautifulsoup4==4.12.2
pillow==10.0.0
tqdm==4.66.1
langchain==0.0.309
speechrecognition
pyttsx3
pyaudio
requests
python-dotenv
anthropic